Add Babergh District Council collector and extract PlacecubeCollectorBase
Closes #564
