+++ BinDays.Api.Collectors/Collectors/Councils/BaberghDistrictCouncil.cs
@@ -1,0 +1,60 @@
+namespace BinDays.Api.Collectors.Collectors.Councils;
+
+using BinDays.Api.Collectors.Collectors.Vendors;
+using BinDays.Api.Collectors.Models;
+using System;
+using System.Collections.Generic;
+
+/// <summary>
+/// Collector implementation for Babergh District Council.
+/// </summary>
+internal sealed class BaberghDistrictCouncil : PlacecubeCollectorBase, ICollector
+{
+	/// <inheritdoc/>
+	public string Name => "Babergh District Council";
+
+	/// <inheritdoc/>
+	public Uri WebsiteUrl => new("https://babergh.gov.uk/");
+
+	/// <inheritdoc/>
+	public override string GovUkId => "babergh";
+
+	/// <inheritdoc/>
+	protected override string BaseUrl => "https://babergh.gov.uk";
+
+	/// <inheritdoc/>
+	protected override IReadOnlyCollection<Bin> BinTypes =>
+	[
+		new()
+		{
+			Name = "General Waste",
+			Colour = BinColour.Black,
+			Keys = [ "Refuse Collection (General Rubbish)" ],
+		},
+		new()
+		{
+			Name = "Recycling",
+			Colour = BinColour.Blue,
+			Keys = [ "Recycling Collection" ],
+		},
+		new()
+		{
+			Name = "Paper and Card Recycling",
+			Colour = BinColour.Green,
+			Keys = [ "Paper And Card Collection" ],
+		},
+		new()
+		{
+			Name = "Food Waste",
+			Colour = BinColour.Grey,
+			Keys = [ "Food Waste Collection" ],
+			Type = BinType.Caddy,
+		},
+		new()
+		{
+			Name = "Garden Waste",
+			Colour = BinColour.Brown,
+			Keys = [ "Garden Waste Collection (Brown Bin)" ],
+		},
+	];
+}
