+++ BinDays.Api.IntegrationTests/Collectors/Councils/BaberghDistrictCouncilTests.cs
@@ -1,0 +1,32 @@
+namespace BinDays.Api.IntegrationTests.Collectors.Councils;
+
+using BinDays.Api.Collectors.Collectors.Councils;
+using BinDays.Api.IntegrationTests.Helpers;
+using System.Threading.Tasks;
+using Xunit;
+using Xunit.Abstractions;
+
+public class BaberghDistrictCouncilTests
+{
+	private readonly IntegrationTestClient _client;
+	private readonly ITestOutputHelper _outputHelper;
+	private static readonly string _govUkId = new BaberghDistrictCouncil().GovUkId;
+
+	public BaberghDistrictCouncilTests(ITestOutputHelper outputHelper)
+	{
+		_outputHelper = outputHelper;
+		_client = new IntegrationTestClient(outputHelper);
+	}
+
+	[Theory]
+	[InlineData("CO10 0AE")]
+	public async Task GetBinDaysTest(string postcode)
+	{
+		await TestSteps.EndToEnd(
+			_client,
+			postcode,
+			_govUkId,
+			_outputHelper
+		);
+	}
+}
